refactor(tests): use alias import in dynamic require for biasReportRoutes
